requirements.txt
flask
discord.py
requests
PyNaCl
python-dotenv